Add files via upload
This is the thesis I wrote for my bachelor's degree in astronomy. It is a review of different topics concerning neutron stars (NS) and their equation of state.
The very basic properties of neutron stars are summarized, as well as a brief description of the very first model for their interiors. Then, the event GW170817 is described and the implications of that in constraining the NS's equation of state are reported.